+++ lib/src/di/injection_container.dart
@@ -1,4 +1,5 @@
 import 'package:get_it/get_it.dart';
+
 import '../services/navigation_history_service.dart';
 
 /// Service locator instance.