+++ lib/src/utils/enums.dart
@@ -1,0 +1,17 @@
+/// Layout options for the navigation stack preview.
+enum StackPreviewLayout {
+  /// A horizontal list of thumbnails.
+  carousel,
+
+  /// A vertical list without thumbnails
+  list,
+}
+
+/// Position options for the navigation stack preview panel.
+enum StackPreviewPosition {
+  /// Panel slides in from the top.
+  top,
+
+  /// Panel slides in from the bottom.
+  bottom,
+}